fix with boot CommandLineRunner
